Remove untextured "mesh_0" from gltf
Vuforia's authoring mesh includes an untextured mesh for places that it
has only geometry data. Removes this untextured mesh from visuals and
collision detection to make the mesh look better and keep collision
detection accurate. This mesh could be added to the BVH to make it part
of the collision detection while invisible but I didn't do that so it
would be visually consistent.